bot/app/requirements.txt
PySide6
trimesh
numpy
pythonocc-core
